+++ packages/micropip/test_micropip.py
@@ -228,3 +228,18 @@ def test_install_different_version2(selenium_standalone_micropip):
         `);
         """
     )
+
+
+@pytest.mark.parametrize("jinja2", ["jinja2", "Jinja2"])
+def test_install_mixed_case2(selenium_standalone_micropip, jinja2):
+    selenium = selenium_standalone_micropip
+    selenium.run_js(
+        f"""
+        await pyodide.loadPackage("micropip");
+        await pyodide.runPythonAsync(`
+            import micropip
+            await micropip.install("{jinja2}")
+            import jinja2
+        `)
+        """
+    )